+++ CHANGES.md
@@ -17,5 +17,6 @@ Apollo 2.1.0
 * [Allow users to associate multiple public namespaces at a time](https://github.com/apolloconfig/apollo/pull/4437)
 * [Move apollo-demo, scripts/docker-quick-start and scripts/apollo-on-kubernetes out of main repository](https://github.com/apolloconfig/apollo/pull/4440)
 * [Add search key when comparing Configuration items](https://github.com/apolloconfig/apollo/pull/4459)
+* [A user-friendly user management page for apollo portal](https://github.com/apolloconfig/apollo/pull/4464)
 ------------------
 All issues and pull requests are [here](https://github.com/apolloconfig/apollo/milestone/11?closed=1)